+++ src/main/kotlin/net/pwall/yaml/YAMLException.kt
@@ -4,4 +4,4 @@ import net.pwall.json.JSONException
 import net.pwall.yaml.parser.Line
 
 class YAMLException(message: String, line: Line, e: Exception? = null) :
-        JSONException("$message at ${line.lineNumber}:${line.index}", e)
+        JSONException("$message at ${line.lineNumber}:${line.index + 1}", e)